worker

+++ src/main/java/ru/pankov/store/service/ProductServiceImpl.java
@@ -3,12 +3,13 @@ package ru.pankov.store.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import ru.pankov.store.dao.ProductRepository;
 import ru.pankov.store.dto.ProductDTO;
 import ru.pankov.store.entity.Product;
+import ru.pankov.store.service.inter.ProductService;
 
-import java.math.BigDecimal;
 import java.util.Optional;
 
 @Service
@@ -18,8 +19,8 @@ public class ProductServiceImpl implements ProductService {
     private final ProductRepository productRepository;
 
     @Override
-    public Page<ProductDTO> findAll(BigDecimal min, BigDecimal max, Integer page, Integer limit) {
-        Page<Product> productsPage = productRepository.findAllByPriceBetween(min, max, PageRequest.of(page - 1, limit));
+    public Page<ProductDTO> findAll(Specification<Product> spec, Integer page, Integer limit) {
+        Page<Product> productsPage = productRepository.findAll(spec, PageRequest.of(page - 1, limit));
         return productsPage.map(ProductDTO::new);
     }
 
@@ -29,23 +30,37 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public void save(ProductDTO productDTO) {
+    public boolean save(ProductDTO productDTO) {
         Product p = new Product(productDTO);
         if (p.getId() != null) {
-            p.setCreatedAt(productRepository.findById(p.getId()).orElseThrow(() -> new RuntimeException("Invalid product")).getCreatedAt());
+            Optional<Product> found = productRepository.findById(p.getId());
+            if (!found.isPresent()) {
+                return false;
+            }
+            p.setCreatedAt(found.get().getCreatedAt());
         }
         productRepository.save(p);
+        return true;
     }
 
     @Override
-    public void deleteById(Long id) {
-        productRepository.deleteById(id);
+    public boolean deleteById(Long id) {
+        if (productRepository.findById(id).isPresent()) {
+            productRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     @Override
-    public void delete(ProductDTO productDTO) {
-        productRepository.delete(productRepository.findById(productDTO.getId()).orElseThrow(() -> new RuntimeException("Invalid product")));
+    public boolean delete(ProductDTO productDTO) {
+        Optional<Product> p = productRepository.findById(productDTO.getId());
+        if (p.isPresent()) {
+            productRepository.delete(p.get());
+            return true;
+        } else {
+            return false;
+        }
     }
-
-
 }
